src/bilanciafrigo-bot/requirements.txt

python-telegram-bot==13.15
paho-mqtt==1.6.1
